refactor(ui): consolidate settings panel to 4 groups and drop dead stubs
Collapse the settings IA from 8 groups / 20 pages to 4 groups / 15 pages:

- New Overview page (default landing) merges Health & Stats with the old
  General page (version identity + telemetry), status panels first and
  degraded services sorted to the top.
- Hardware & Runtimes merges the two read-only evidence pages
  (Backend & GPU + Runtimes) into one surface.
- Loaded Models absorbs the Library & Downloads panels; Updates absorbs
  the About panel.
- Remove the fully blocked Hardware Tuning stub page, the permanently
  gated ROCm/gfx placeholder rows, the dead disabled Pre-pull button,
  and Security's hardcoded "unknown" client-key pip and permanent
  "status unavailable" throttle panel (no backend routes back any of
  them).
- Friendlier naming: Image-gen -> Image Generation, Agents / Brain ->
  Agent Chat.
- Legacy #settings/<id> deep links (general, health, library, backend,
  runtimes, hwtuning, about) resolve through SECTION_ALIASES; command
  palette entries updated to the new tree.

Verified: eslint, tsc, vitest (75), full Playwright suite green.
