+++ ui/src/dash/command-palette.jsx
@@ -275,19 +275,21 @@ function buildCommandItems(slots, models, activePull, owuiUrl = "") {
   });
 
   // Settings sections — deep links into #settings/<section>. Mirrors
-  // VALID_IDS in settings.jsx; Auth/Runtime entries removed (auth was
-  // dropped in ADR-0012 and no Runtime section exists).
+  // VALID_IDS in SettingsNav.jsx (settings-panel cleanup: 4 groups; legacy
+  // ids like general/about/health resolve through SECTION_ALIASES).
   [
-    { id: "set-secrets",   label: "Secrets",       route: "settings/secrets",  sub: "HF_TOKEN, provider keys, custom env vars" },
-    { id: "set-storage",   label: "Storage",       route: "settings/storage",  sub: "[models].store · auto_scan · file extensions" },
-    { id: "set-updates",   label: "Updates",       route: "settings/updates",  sub: "check / install / roll back · channel" },
-    { id: "set-voice",     label: "Voice",         route: "settings/voice",    sub: "STT + TTS models, default voice" },
-    { id: "set-imagegen",  label: "Image-gen",     route: "settings/imagegen", sub: "engine, model, generation defaults" },
-    { id: "set-npu",       label: "NPU",           route: "settings/npu",      sub: "FLM context size, embed/ASR at boot, occupancy" },
-    { id: "set-defaults",  label: "Default slots", route: "settings/defaults", sub: "per-modality default slot" },
-    { id: "set-general",   label: "General",       route: "settings/general",  sub: "telemetry, appearance" },
-    { id: "set-advanced",  label: "Advanced",      route: "settings/advanced", sub: "slots runtime, dispatcher, memory, activity · restart hal0-api" },
-    { id: "set-about",     label: "About",         route: "settings/about",    sub: "version, license, links" },
+    { id: "set-overview",  label: "Overview",             route: "settings/overview", sub: "health, services, hardware, requests · telemetry, version" },
+    { id: "set-security",  label: "Security",             route: "settings/security", sub: "auth enforcement, key rotation, route exposure" },
+    { id: "set-doctor",    label: "Doctor",               route: "settings/doctor",   sub: "system diagnostics" },
+    { id: "set-defaults",  label: "Loaded Models",        route: "settings/slots",    sub: "per-modality default slot · slot pool · library" },
+    { id: "set-voice",     label: "Voice",                route: "settings/voice",    sub: "STT + TTS models, default voice" },
+    { id: "set-imagegen",  label: "Image Generation",     route: "settings/imagegen", sub: "engine, model, generation defaults" },
+    { id: "set-npu",       label: "NPU",                  route: "settings/npu",      sub: "FLM context size, embed/ASR at boot, occupancy" },
+    { id: "set-hardware",  label: "Hardware & Runtimes",  route: "settings/hardware", sub: "detected hardware, backends, runner images" },
+    { id: "set-storage",   label: "Storage",              route: "settings/storage",  sub: "[models].store · auto_scan · file extensions" },
+    { id: "set-updates",   label: "Updates",              route: "settings/updates",  sub: "check / install / roll back · channel · about" },
+    { id: "set-advanced",  label: "Advanced",             route: "settings/advanced", sub: "slots runtime, dispatcher, memory, activity · restart hal0-api" },
+    { id: "set-secrets",   label: "Secrets",              route: "settings/secrets",  sub: "HF_TOKEN, provider keys, custom env vars" },
   ].forEach(s => items.push({ ...s, section: "Settings", icon: Icons.settings }));
 
   // Global actions — every one is wired to a real effect.
