+++ ui/src/dash/settings/SettingsNav.jsx
@@ -1,89 +1,62 @@
-// Grouped settings nav (P3-ui split phase 1, spec (c) target tree: SERVER /
-// MODELS / INFERENCE / ROUTING / OBSERVABILITY / DATA / DIAGNOSTICS /
-// INTEGRATIONS). Replaces the old settings.jsx flat 12-item rail with a
-// grouped one; every existing section keeps its original `id` so
-// #settings/<id> deep links are unaffected — only the label/grouping
-// changed for the four sections whose id was itself renamed in the old UI
-// text ("Slots" → "Loaded Models", "Agents / Brain" unchanged).
-//
-// `disabled` entries render but refuse selection — used for the two pages
-// gated on unbuilt backend lanes (Security → auth §1, Hardware Tuning →
-// host-tuning §2/§21.1). Everything else here has a real page module; the
-// stub pages (Library & Downloads, Health & Stats, Doctor) are visible and
-// clickable — they're just not wired to real data yet (see each page's
-// header comment).
+// Grouped settings nav — settings-panel-cleanup consolidation. The old
+// 8-group / 20-page tree collapsed to 4 groups / 15 pages:
+//   - Overview (new, default landing) absorbs Health & Stats + General
+//   - Loaded Models absorbs Library & Downloads
+//   - Hardware & Runtimes merges Backend & GPU + Runtimes (both evidence UIs)
+//   - Updates absorbs About
+//   - Hardware Tuning (blocked stub, spec D4) removed until its lane lands
+// Every retired `id` stays routable via SECTION_ALIASES so existing
+// #settings/<id> deep links land on the page that absorbed the content.
 export const NAV_GROUPS = [
   {
-    title: "SERVER",
+    title: "GENERAL",
     items: [
-      { id: "general", label: "General" },
-      // D4 (post-R3 surface rework): un-disabled — the auth surface landed
-      // (GET /api/auth/status). The page is status-only and flags every
-      // missing affordance (rotation, client-key set, throttle counts,
-      // per-route exposure table) as an API-lane request in-place.
+      { id: "overview", label: "Overview" },
       { id: "security", label: "Security" },
+      { id: "doctor", label: "Doctor" },
     ],
   },
   {
-    title: "MODELS",
+    title: "MODELS & INFERENCE",
     items: [
       { id: "slots", label: "Loaded Models" },
-      { id: "library", label: "Library & Downloads" },
-      // ML-4 landed the runner-image registry + model-config taxonomy, so the
-      // per-model launch-defaults page is now buildable (spec §7.1a/d).
       { id: "modeldefaults", label: "Model Defaults" },
-    ],
-  },
-  {
-    title: "INFERENCE",
-    items: [
-      // ML-4 landed the runner-image registry, unblocking the Backend & GPU
-      // introspection surface (spec §7.1b, detected hw via /api/hardware).
-      { id: "backend", label: "Backend & GPU" },
-      { id: "hwtuning", label: "Hardware Tuning", disabled: true, disabledReason: "coming with the tuning lane (spec §2/§21.1)" },
-      { id: "npu", label: "NPU" },
       { id: "voice", label: "Voice" },
-      { id: "imagegen", label: "Image-gen" },
+      { id: "imagegen", label: "Image Generation" },
+      { id: "npu", label: "NPU" },
     ],
   },
   {
-    title: "ROUTING",
+    title: "SYSTEM",
     items: [
-      { id: "agents", label: "Agents / Brain" },
-    ],
-  },
-  {
-    title: "OBSERVABILITY",
-    items: [
-      { id: "health", label: "Health & Stats" },
-    ],
-  },
-  {
-    title: "DATA",
-    items: [
+      { id: "hardware", label: "Hardware & Runtimes" },
       { id: "storage", label: "Storage" },
       { id: "memory", label: "Memory" },
-    ],
-  },
-  {
-    title: "DIAGNOSTICS",
-    items: [
-      { id: "doctor", label: "Doctor" },
       { id: "updates", label: "Updates" },
-      // D3 (post-R3 surface rework): runner/image evidence page. Grouped here
-      // with Updates as the system & updates axis, far from the model editors.
-      { id: "runtimes", label: "Runtimes" },
       { id: "advanced", label: "Advanced" },
-      { id: "about", label: "About" },
     ],
   },
   {
     title: "INTEGRATIONS",
     items: [
       { id: "secrets", label: "Secrets" },
+      { id: "agents", label: "Agent Chat" },
     ],
   },
 ];
+
+// Legacy section ids → the page that owns that content now. Applied before
+// the VALID_IDS check so old bookmarks, palette entries, and cross-links keep
+// resolving.
+export const SECTION_ALIASES = {
+  general: "overview",
+  health: "overview",
+  library: "slots",
+  backend: "hardware",
+  runtimes: "hardware",
+  hwtuning: "hardware",
+  about: "updates",
+};
 
 // Flat id → item lookup + the ordered id list, derived once from NAV_GROUPS
 // so the shell's VALID_IDS check and the nav's render loop can't drift.
@@ -103,13 +76,10 @@ export function SettingsNav({ section, onSelect }) {
           {group.items.map(item => (
             <div
               key={item.id}
-              className={"nav-item" + (section === item.id ? " active" : "") + (item.disabled ? " disabled" : "")}
-              title={item.disabled ? item.disabledReason : undefined}
-              onClick={() => { if (!item.disabled) onSelect(item.id); }}
-              style={item.disabled ? {opacity: 0.5, cursor: "not-allowed"} : undefined}
+              className={"nav-item" + (section === item.id ? " active" : "")}
+              onClick={() => onSelect(item.id)}
             >
               {item.label}
-              {item.disabled && <span style={{marginLeft: 6}}>{item.id === "hwtuning" ? "🔒" : "⛔"}</span>}
             </div>
           ))}
         </div>
