+++ ui/tests/e2e/specs/runtimes-page-v3.spec.ts
@@ -1,11 +1,13 @@
 /**
- * runtimes-page-v3 — D3 Settings → Runtimes (post-R3 surface rework).
+ * runtimes-page-v3 — runner-image evidence, now on Settings → Hardware &
+ * Runtimes (settings-panel cleanup merged the Runtimes page into it).
  *
  * Evidence UI for the runner/image axis: one row per RUNNER_IMAGES entry with
  * family, image ref + digest (read-only), on-disk state, and the models/slots
- * reverse index. Nothing on the page edits an image string; the pull affordance
- * is disabled-with-reason (no per-runner pull endpoint), and a fully-unavailable
- * probe (no podman) degrades gracefully with a reason.
+ * reverse index. Nothing on the page edits an image string; the old disabled
+ * "Pre-pull" placeholder is gone (no per-runner pull endpoint — a dead
+ * control is worse than none), and a fully-unavailable probe (no podman)
+ * degrades gracefully with a reason.
  *
  * /api/system-info is NOT in the in-bundle mock allowlist, so page.route wins.
  */
@@ -23,11 +25,11 @@ function mockSystemInfo(page: import('@playwright/test').Page, backends: Record<
 
 async function openRuntimes(page: import('@playwright/test').Page) {
   await page.goto('/#settings')
-  await page.locator('.settings-nav .nav-item', { hasText: 'Runtimes' }).click()
-  await expect(page.getByTestId('runtimes-page')).toBeVisible()
+  await page.locator('.settings-nav .nav-item', { hasText: 'Hardware & Runtimes' }).click()
+  await expect(page.getByTestId('hardware-page')).toBeVisible()
 }
 
-test.describe('Settings → Runtimes', () => {
+test.describe('Settings → Hardware & Runtimes (runner images)', () => {
   test('renders runner rows with status derived from system-info state', async ({ page }) => {
     await mockSystemInfo(page, {
       rocmfpx: { image: 'ghcr.io/hal0ai/hal0-rocmfp4:v0.9.4', runtime_family: 'llama-server', device_class: 'gpu', backend: 'rocm', state: 'installed' },
@@ -46,10 +48,10 @@ test.describe('Settings → Runtimes', () => {
     await expect(page.getByTestId('runtime-image-rocmfpx')).toContainText('hal0-rocmfp4')
 
     // Evidence UI: nothing on the page edits an image string.
-    await expect(page.getByTestId('runtimes-page').locator('input, textarea')).toHaveCount(0)
+    await expect(page.getByTestId('hardware-page').locator('input, textarea')).toHaveCount(0)
 
-    // The not-pulled runner offers a pre-pull, but it is disabled (no endpoint).
-    await expect(page.getByTestId('runtime-prepull-vulkanfpx')).toBeDisabled()
+    // The dead disabled "Pre-pull" placeholder is gone (no endpoint backs it).
+    await expect(page.getByTestId('runtime-prepull-vulkanfpx')).toHaveCount(0)
   })
 
   test('degraded probe (all unavailable) shows a reason', async ({ page }) => {
